% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_water_surface.R
\name{classify_water_surface}
\alias{classify_water_surface}
\title{Title}
\usage{
classify_water_surface(s1_data, rf_model, output_dir = "./RESULTS/Test")
}
\arguments{
\item{s1_data}{}

\item{rf_model}{}

\item{output_dir}{}
}
\description{
Title
}
