% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_data.R
\name{preprocessing_data}
\alias{preprocessing_data}
\title{Title}
\usage{
preprocessing_data(stack.radar, aligned, index, stations, timestep)
}
\arguments{
\item{stack.radar}{}

\item{aligned}{}

\item{index}{}

\item{stations}{}

\item{timestep}{This parameter is .....}
}
\description{
Title
}
