% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train_rf.R
\name{train_rf_model}
\alias{train_rf_model}
\title{Title}
\usage{
train_rf_model(training_data, ntree = 500, por.train = 0.7)
}
\arguments{
\item{training_data}{}

\item{ntree}{}

\item{por.train}{}
}
\description{
Title
}
