% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_index_stack.R
\name{radar_index_stack}
\alias{radar_index_stack}
\title{Title}
\usage{
radar_index_stack(VV, VH, index, nameID)
}
\arguments{
<<<<<<< HEAD
\item{VV}{}
=======
\item{VV}{Is the backscatter of the signal transmited and recived vertical}
>>>>>>> 991fa7e67d056a74acea2130cac016f0099c329d

\item{VH}{}

\item{index}{}

\item{nameID}{}
}
\description{
Title
}
